Move WalletLocalData to its own file
Also give the class a better name.
